requirements.txt
langchain
langgraph
langsmith
langchain-community
langchain-qdrant
langchain-huggingface
qdrant-client
sentence-transformers
streamlit
pypdf
python-dotenv
requests
langchain-google-genai
pytest
pytest-asyncio
typing-extensions
tenacity

